requirements.txt
flask>=3.1.0
flask-cors>=6.0.0
python-docx>=1.1.0
markdown>=3.6
spacy>=3.8.7
sentence-transformers>=3.4.1
numpy>=2.0.0
scipy>=1.13.0
scikit-learn>=1.5.0
pyphen>=0.14.0
chardet>=5.0.0
